pyproject.toml
[project]
name = "story-lord"
version = "1.0.0"
description = "Default template for PDM package"
authors = [{ name = "Koen van Esterik", email = "koen@vanesterik.com" }]
dependencies = [
    "openai>=1.8.0",
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.4",
    "types-toml>=0.10.8.7",
]
